Remove the dev folder and add musicxml_inspector folder and create sub-folder for controller, model, and widgets to organized the structure of the code.
